+++ lib/offers/infrastructure/models/salary_model.dart
@@ -2,7 +2,7 @@ import 'package:teacher_finder/offers/domain/entities/enums/currency.dart';
 import 'package:teacher_finder/offers/domain/entities/salary.dart';
 
 class SalaryModel extends Salary {
-  SalaryModel({required int mount, required Currency currency})
+  SalaryModel({required double mount, required Currency currency})
       : super(mount: mount, currency: currency);
 
   factory SalaryModel.fromJson(Map<String, dynamic> json) {
